pyproject.toml
[project]
name = "veoci-solution-mapper"
version = "0.1.0"
description = "CLI tool to map Veoci solution structure and dependencies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "networkx>=3.0",
    "questionary>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.56.0",
]

[project.scripts]
veoci-map = "veoci_mapper.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/veoci_mapper"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
